backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
openai==0.28.1
python-dotenv==1.0.0
stripe==12.0.0
python-multipart
pillow
httpx
huggingface_hub>=0.19.0
